Q2/requirements.txt
requests
Flask
python-dotenv 